backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
python-dotenv
google-genai
slowapi
